Re-designed Mail package to be much simpeler.
Re-designed Mail interface (renamed to Mailer) to send a simple data object instead of doing all the logic in the mailer implementing this interface. This is because the methods (now in the data object) never changes but different methods of sending the email have been made. So setting it up this way we need to implement only 1 very simple method and can do only the sending logic in the class implementing this interface.